refactor(ai-chat): decompose AIChatViewModel into responsibility extensions
