+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - AI Chat: composer focus ring uses the standard macOS accent stroke instead of a colored gradient
 - AI inline suggestions: debounce now uses structured Swift concurrency, and the delay is configurable via the `inlineSuggestionDebounceMs` setting (default 500ms)
 - Copilot LSP shutdown caps at 10 seconds, closes pipes explicitly, and strips the quarantine attribute from the downloaded binary
+- AI Chat: streaming view model split into focused extensions backed by a single `streamingState` enum
 
 ## [0.39.1] - 2026-05-08
