+++ TablePro/Core/AI/String+AIEndpoint.swift
@@ -1,0 +1,12 @@
+//
+//  String+AIEndpoint.swift
+//  TablePro
+//
+
+import Foundation
+
+extension String {
+    func normalizedEndpoint() -> String {
+        hasSuffix("/") ? String(dropLast()) : self
+    }
+}
